Moved to client.rs and added docs.
