+++ src/client.rs
@@ -1,12 +1,18 @@
 use std::collections::HashMap;
 
 use async_trait::async_trait;
+use futures::StreamExt;
 use prost_types::Any;
 use serde::{Deserialize, Serialize};
-use tonic::Streaming;
-use tonic::{transport::Channel as TonicChannel, Request};
+use tonic::{Request, transport::Channel as TonicChannel};
+use tonic::{Status, Streaming};
+use tonic::codegen::tokio_stream;
 
 use crate::dapr::dapr::proto::{common::v1 as common_v1, runtime::v1 as dapr_v1};
+use crate::dapr::dapr::proto::common::v1::StreamPayload;
+use crate::dapr::dapr::proto::runtime::v1::{
+    DecryptRequest, DecryptRequestOptions, EncryptRequest, EncryptRequestOptions,
+};
 use crate::error::Error;
 
 #[derive(Clone)]
@@ -625,5 +631,117 @@ where
             value,
             ..Default::default()
         }
+    }
+}
+
+impl Client<TonicClient> {
+    /// Encrypt binary data using Dapr. returns Vec<StreamPayload> to be used in decrypt method
+    ///
+    /// # Arguments
+    ///
+    /// * `payload` - A Reference to the data to encrypt, should impl Into<Vec<u8>>
+    /// * `request_option` - Encryption request options.
+    ///
+    /// # Example
+    /// ```
+    /// let encrypted = client
+    ///         .encrypt(
+    ///             &"Test".to_string(),
+    ///             EncryptRequestOptions {
+    ///                 component_name: "localstorage".to_string(),
+    ///                 key_name: "rsa-private-key.pem".to_string(),
+    ///                 key_wrap_algorithm: "RSA".to_string(),
+    ///                 data_encryption_cipher: "aes-gcm".to_string(),
+    ///                 omit_decryption_key_name: false,
+    ///                 decryption_key_name: "rsa-private-key.pem".to_string(),
+    ///             },
+    ///         )
+    ///         .await
+    ///         .unwrap();
+    /// ```
+    pub async fn encrypt<T>(
+        &mut self,
+        payload: &T,
+        request_options: EncryptRequestOptions,
+    ) -> Result<Vec<StreamPayload>, Status>
+    where
+        T: Into<Vec<u8>> + Clone,
+    {
+        let stream_payload = StreamPayload {
+            data: payload.clone().into(),
+            seq: 0,
+        };
+        let request = EncryptRequest {
+            options: Some(request_options),
+            payload: Some(stream_payload),
+        };
+        let request = Request::new(tokio_stream::iter([request]));
+        let stream = self.0.encrypt_alpha1(request).await?;
+        let mut stream = stream.into_inner();
+        let mut return_data = vec![];
+        while let Some(resp) = stream.next().await {
+            if let Ok(resp) = resp {
+                if let Some(data) = resp.payload {
+                    return_data.push(data)
+                }
+            }
+        }
+        Ok(return_data)
+    }
+
+    /// Decrypt binary data using Dapr. returns Vec<u8>.
+    ///
+    /// # Arguments
+    ///
+    /// * `encrypted` - Encrypted data usually returned from encrypted, Vec<StreamPayload>
+    /// * `options` - Decryption request options.
+    ///
+    /// # Example
+    /// ```
+    /// let decrypted = client
+    ///         .decrypt(
+    ///             encrypted,
+    ///             DecryptRequestOptions {
+    ///                 component_name: "localstorage".to_string(),
+    ///                 key_name: "rsa-private-key.pem".to_string(),
+    ///             },
+    ///         )
+    ///         .await
+    ///         .unwrap();
+    /// ```
+    pub async fn decrypt(
+        &mut self,
+        encrypted: Vec<StreamPayload>,
+        options: DecryptRequestOptions,
+    ) -> Result<Vec<u8>, Status> {
+        let requested_items: Vec<DecryptRequest> = encrypted
+            .iter()
+            .enumerate()
+            .map(|(i, item)| {
+                if i == 0 {
+                    DecryptRequest {
+                        options: Some(options.clone()),
+                        payload: Some(item.clone()),
+                    }
+                } else {
+                    DecryptRequest {
+                        options: None,
+                        payload: Some(item.clone()),
+                    }
+                }
+            })
+            .collect();
+        let request = Request::new(tokio_stream::iter(requested_items));
+        let stream = self.0.decrypt_alpha1(request).await?;
+        let mut stream = stream.into_inner();
+        let mut data = vec![];
+        while let Some(resp) = stream.next().await {
+            if let Ok(resp) = resp {
+                if let Some(mut payload) = resp.payload {
+                    data.append(payload.data.as_mut())
+                }
+            }
+        }
+        Ok(data)
     }
 }
