+++ src/lib.rs
@@ -1,15 +1,13 @@
-pub mod appcallback;
-pub mod client;
-pub mod crypto;
-pub mod dapr;
-pub mod error;
-pub mod server;
+extern crate dapr_macros;
 
+pub use dapr_macros::actor;
 pub use serde;
-
 pub use serde_json;
 
 pub use client::Client;
 
-extern crate dapr_macros;
-pub use dapr_macros::actor;
+pub mod appcallback;
+pub mod client;
+pub mod dapr;
+pub mod error;
+pub mod server;
